Add minimum index value to castra
Currently we record the maximum value of each incoming partition.
Assuming that partitions come in sequentially this gives us nice
min/max ranges on all partitions except the first, which is unbounded.

This adds a `.minimum` field to Castra to fill this place.

This feels a bit kludgy but does solve an immediate problem.
Alternative solutions welcome.

Knowing the minimum ends up being very useful when two tables interact.